refactor: Wings 구현체에서의 중복을 제거
기존에는 Wings 추상 클래스 하위에 HanWings와 ChoWings 구현체가 있었다.

하지만 두 구현체에서 수행하는 로직이 거의 동일하여 제대로 추상화되지 않는 문제가 있었다.

이를 해결하기 위해 Wings를 추상 클래스에서 일반 클래스로 변경한다.
